refactor(ui): extract diagnostics/render/sync-status.ts — renderSyncStatus
Move renderSyncStatus into its own module. The module owns the grid
entries (state, pending events, last sync, last ping, retention), the
meta-line composition, and the daemon-state-driven action list. The
barrel re-exports renderSyncStatus so existing call sites stay on the
same import path.
